# openapi.yaml
openapi: 3.0.3
info:
  title: Rental Management API
  description: |
    A comprehensive API for managing rental properties, units, expenses, and income.

    This API provides endpoints for:
    - Managing rental units (CRUD operations)
    - Managing expenses (CRUD operations with filtering)
    - Managing income (CRUD operations with filtering)
    - Expense and income statistics and reporting

    All data is persisted using Upstash Redis KV storage with automatic fallback to localStorage during development.
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://your-app.vercel.app
    description: Production server
paths:
  /api/units:
    get:
      summary: Get all units
      description: Retrieve a list of all rental units
      tags:
        - Units
      responses:
        "200":
          description: Successfully retrieved units
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Unit"
              example:
                - id: "unit-1"
                  name: "Commercial Store"
                  property: "Commercial"
                  location: "Downtown"
                  address: "123 Main St"
                  monthlyRent: 2500
                  tenant:
                    name: "John's Coffee Shop"
                    phone: "+1-555-0123"
                    email: "john@coffee.com"
                  contractStart: "2024-01-01"
                  contractEnd: "2024-12-31"
        "500":
          $ref: "#/components/responses/ServerError"

    post:
      summary: Create a new unit
      description: Create a new rental unit
      tags:
        - Units
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Unit"
            example:
              id: "unit-2"
              name: "House A - Unit 1"
              property: "Residential"
              location: "Northside"
              address: "456 Oak Ave"
              monthlyRent: 1800
              tenant:
                name: "Jane Smith"
                phone: "+1-555-0456"
                email: "jane@email.com"
              contractStart: "2024-02-01"
      responses:
        "201":
          description: Unit created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Unit"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/ServerError"
  /api/units/{id}:
    get:
      summary: Get a specific unit
      description: Retrieve a single unit by its ID
      tags:
        - Units
      parameters:
        - name: id
          in: path
          required: true
          description: The unique identifier of the unit
          schema:
            type: string
          example: "unit-1"
      responses:
        "200":
          description: Successfully retrieved unit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Unit"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/ServerError"

    put:
      summary: Update a specific unit
      description: Update an existing unit's information
      tags:
        - Units
      parameters:
        - name: id
          in: path
          required: true
          description: The unique identifier of the unit
          schema:
            type: string
          example: "unit-1"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UnitUpdate"
            example:
              monthlyRent: 2600
              tenant:
                name: "John's Updated Coffee Shop"
                phone: "+1-555-0124"
      responses:
        "200":
          description: Unit updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Unit"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/ServerError"

    delete:
      summary: Delete a specific unit
      description: Remove a unit from the system
      tags:
        - Units
      parameters:
        - name: id
          in: path
          required: true
          description: The unique identifier of the unit
          schema:
            type: string
          example: "unit-1"
      responses:
        "200":
          description: Unit deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/ServerError"
...
components:
  schemas:
    Tenant:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          description: Full name of the tenant
          example: "John Smith"
        phone:
          type: string
          description: Contact phone number
          example: "+1-555-0123"
        email:
          type: string
          format: email
          description: Contact email address
          example: "john@email.com"
    Unit:
      type: object
      required:
        - id
        - name
        - property
        - location
        - address
        - monthlyRent
        - tenant
        - contractStart
      properties:
        id:
          type: string
          description: Unique identifier for the unit
          example: "unit-1"
        name:
          type: string
          description: Display name of the unit
          example: "Commercial Store"
        property:
          type: string
          enum: [Commercial, Residential]
          description: Type of property
          example: "Commercial"
        location:
          type: string
          enum: [Downtown, Northside, Southside]
          description: Location area
          example: "Downtown"
        address:
          type: string
          description: Physical address of the unit
          example: "123 Main St"
        monthlyRent:
          type: number
          description: Monthly rent amount
          minimum: 0
          example: 2500
        tenant:
          $ref: "#/components/schemas/Tenant"
        contractStart:
          type: string
          format: date
          description: Contract start date (ISO 8601)
          example: "2024-01-01"
        contractEnd:
          type: string
          format: date
          description: Contract end date (ISO 8601)
          example: "2024-12-31"
    UnitUpdate:
      type: object
      properties:
        name:
          type: string
          description: Display name of the unit
        property:
          type: string
          enum: [Commercial, Residential]
          description: Type of property
        location:
          type: string
          enum: [Downtown, Northside, Southside]
          description: Location area
        address:
          type: string
          description: Physical address of the unit
        monthlyRent:
          type: number
          description: Monthly rent amount
          minimum: 0
        tenant:
          $ref: "#/components/schemas/Tenant"
        contractStart:
          type: string
          format: date
          description: Contract start date (ISO 8601)
        contractEnd:
          type: string
          format: date
          description: Contract end date (ISO 8601)
...
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: Error message
          example: "Resource not found"
  responses:
    BadRequest:
      description: Bad request - Invalid input data
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: "Missing required field: name"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: "Unit not found"
    ServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: "Internal server error"
...